Created MavericksView
